% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5_graph.R
\name{as_igraph}
\alias{as_igraph}
\title{Convert a Step 5 graph object to igraph}
\usage{
as_igraph(x, ...)
}
\arguments{
\item{x}{A \code{"gllrm_graph"} object.}

\item{...}{Ignored.}
}
\value{
An \code{igraph} object.
}
\description{
Convert a Step 5 graph object to igraph
}
