% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_checks.R
\name{as_screening_data_frame}
\alias{as_screening_data_frame}
\title{Coerce input to data frame for screening functions}
\usage{
as_screening_data_frame(dataset, arg_name = "dataset")
}
\arguments{
\item{dataset}{A tabular object.}

\item{arg_name}{Name of the argument being checked.}
}
\value{
A base data frame.
}
\description{
Coerce input to data frame for screening functions
}
\keyword{internal}
