% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_helper_functions.R
\name{check_covariate}
\alias{check_covariate}
\title{Check covaraite type}
\usage{
check_covariate(x, name)
}
\arguments{
\item{x}{A covariate vector}

\item{name}{name of the covariate, as a string. (only used in case of error)}
}
\value{
If covariate is numeric, returns origional covariate. If covariate
is a factor, it will be converted to numeric starting from 0. If neither,
the function stops and returns cause, that covariate is neither numeric
nor a factor
}
\description{
Check covaraite type
}
\examples{
check_covariate(c(20,22,45), "age")

check_covariate(factor(c("f", "f", "m")), "sex")

}
\keyword{internal}
