% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_helper_functions.R
\name{check_M3_covariate}
\alias{check_M3_covariate}
\title{Check M3 covariate, checks if the direction of association for a covaraite
and score pair is the same for that covariate and each individual item.}
\usage{
check_M3_covariate(df)
}
\arguments{
\item{df}{a dataframe containing the associations, i.e. basically
the output of M3.}
}
\value{
a boolean indocating if the direction of associations are consistent
}
\description{
Check M3 covariate, checks if the direction of association for a covaraite
and score pair is the same for that covariate and each individual item.
}
\keyword{internal}
