% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{combine_step3bc}
\alias{combine_step3bc}
\title{Combine Step 3(b) and 3(c) results}
\usage{
combine_step3bc(
  step3b_results,
  step3c_results,
  original_source_list,
  original_dif_list
)
}
\arguments{
\item{step3b_results}{output from run_step3b}

\item{step3c_results}{output from run_step3c}

\item{original_source_list}{original SOURCE list from s.d_list}

\item{original_dif_list}{original DIF list from s.d_list}
}
\value{
updated SOURCE and DIF lists
}
\description{
Combine Step 3(b) and 3(c) results
}
