% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2_helper_functions.R
\name{compute_and_include_total_score}
\alias{compute_and_include_total_score}
\title{Compute and insert total item score}
\usage{
compute_and_include_total_score(data, y)
}
\arguments{
\item{data}{item repose data}

\item{y}{character vector of item names}
}
\value{
data frame of item response with one additional column containing the total score
}
\description{
Compute and insert total item score
}
\examples{
df <- toy_spadi_pain
items <- paste0("item", 1:5)

## note to self: should we include a way to handle NA's?
}
