% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{compute_partial_gamma}
\alias{compute_partial_gamma}
\title{Partial gamma coefficient}
\usage{
compute_partial_gamma(dataset, Yi, Xj, strata_vars, return_stats = FALSE)
}
\arguments{
\item{dataset}{A data frame containng the data set}

\item{Yi}{The item in question for the hypothesis}

\item{Xj}{The vocaraite in question for the hypothesis}

\item{strata_vars}{the variables we want to condition on and hence stratify with
respect to}

\item{return_stats}{Logical. If \code{TRUE}, return partial gamma together
with its asymptotic standard error, confidence limits, and p-value.}
}
\value{
The partial gamma coefficient corresponding to the test. If
\code{return_stats = TRUE}, returns a list with partial gamma, standard
error, confidence limits, and asymptotic p-value.
}
\description{
Partial gamma coefficient
}
\keyword{internal}
