% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{compute_partial_gamma_stats}
\alias{compute_partial_gamma_stats}
\title{Partial gamma coefficient and asymptotic standard error}
\usage{
compute_partial_gamma_stats(dataset, Yi, Xj, strata_vars, conf.level = 0.95)
}
\arguments{
\item{dataset}{A data frame containing the data set.}

\item{Yi}{The item in question for the hypothesis.}

\item{Xj}{The covariate in question for the hypothesis.}

\item{strata_vars}{Variables defining the conditioning strata.}

\item{conf.level}{Confidence level for confidence limits.}
}
\value{
A list with partial gamma, standard error, confidence limits, and
iarm-style asymptotic p-value.
}
\description{
Partial gamma coefficient and asymptotic standard error
}
\keyword{internal}
