% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_helper_functions.R
\name{compute_total_score}
\alias{compute_total_score}
\title{Total score computer. Computes the total score of each participant, i.e.
the row sum of items per participant}
\usage{
compute_total_score(items_df)
}
\arguments{
\item{items_df}{the sub df only containing the items}
}
\value{
the row sums of the items data frame
}
\description{
Total score computer. Computes the total score of each participant, i.e.
the row sum of items per participant
}
\examples{
compute_total_score(data.frame(item1 = c(1,1,1), item2 = c(2,2,2)))
}
\keyword{internal}
