% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_helper_functions.R
\name{cor_one_pair}
\alias{cor_one_pair}
\title{Compute correlation between two variables. Optionally includes the p-value
from the correlation test.}
\usage{
cor_one_pair(x, y, method = "gamma")
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector of same length as x (variable 1)}

\item{method}{Character string for deciding which correlation method to use.}
}
\value{
A named list with elements:
\describe{
\item{correlation}{Estimated correlation coefficient.}
\item{p_value}{P-value for the test of zero correlation, or \code{NA}.}
}
}
\description{
Compute correlation between two variables. Optionally includes the p-value
from the correlation test.
}
\keyword{internal}
