% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_checks.R
\name{correct.dataset.structure}
\alias{correct.dataset.structure}
\title{convert data set to correct data type}
\usage{
correct.dataset.structure(dataset)
}
\arguments{
\item{dataset}{}
}
\description{
convert data set to correct data type
}
