% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{genuine_LD}
\alias{genuine_LD}
\title{Genuine LD Identifier.}
\usage{
genuine_LD(screen_LD_output, crit_val = 0.05)
}
\arguments{
\item{screen_LD_output}{the output from the screen_LD function.}

\item{crit_val}{The value at which we determine an adjusted pvalue is
critical}
}
\value{
description
}
\description{
This function effectivly performs step 3a. of the posed item screening.
We take the output from the screen_LD function, which flags all LD with a
significant partial gamma coefficient. On this LD data we compute the mean
partial gamma coefficient for each pair (i.e. the mean of the partial gamma
produced by checking the hypothesies \eqn{Y_i \bot Y_j \mid R_i} and
\eqn{Y_j \bot Y_i \mid R_j})
}
