% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{genuine_LD2}
\alias{genuine_LD2}
\title{Genuine LD identifyier}
\usage{
genuine_LD2(
  screen_LD_output,
  number_of_multiple_tests,
  method = "BH",
  crit_val = 0.05
)
}
\arguments{
\item{crit_val}{}
}
\description{
Genuine LD identifyier
}
