% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-llraDat1.R
\docType{data}
\name{llraDat1}
\alias{llraDat1}
\title{Artificial LLRA Dataset}
\format{
A data frame with XXX rows and XXX variables:
\describe{
\item{t1.I1}{Description}
\item{t1.I2}{Description}
\item{t1.I3}{Description}
\item{t1.I4}{Description}
\item{t1.I5}{Description}
\item{t2.I1}{Description}
\item{t2.I2}{Description}
\item{t2.I3}{Description}
\item{t2.I4}{Description}
\item{t2.I5}{Description}
\item{t3.I1}{Description}
\item{t3.I2}{Description}
\item{t3.I3}{Description}
\item{t3.I4}{Description}
\item{t3.I5}{Description}
\item{t4.I1}{Description}
\item{t4.I2}{Description}
\item{t4.I3}{Description}
\item{t4.I4}{Description}
\item{t4.I5}{Description}
\item{t5.I1}{Description}
\item{t5.I2}{Description}
\item{t5.I3}{Description}
\item{t5.I4}{Description}
\item{t5.I5}{Description}
\item{groups}{Description}
}
}
\source{
Adapted from the \pkg{eRm} package.
}
\usage{
llraDat1
}
\description{
An artificial dataset originally provided with the \pkg{eRm} package,
}
\keyword{datasets}
