% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lltmdat1.R
\docType{data}
\name{lltmdat1}
\alias{lltmdat1}
\title{Artificial LLTM Dataset}
\format{
A data frame with XXX rows and XXX variables:
\describe{
\item{I1.t1}{Description}
\item{I2.t1}{Description}
\item{I3.t1}{Description}
\item{I4.t1}{Description}
\item{I5.t1}{Description}
\item{I6.t1}{Description}
\item{I7.t1}{Description}
\item{I8.t1}{Description}
\item{I9.t1}{Description}
\item{I10.t1}{Description}
\item{I11.t1}{Description}
\item{I12.t1}{Description}
\item{I13.t1}{Description}
\item{I14.t1}{Description}
\item{I15.t1}{Description}
\item{I14.t1}{Description}
\item{I1.t2}{Description}
\item{I2.t2}{Description}
\item{I3.t2}{Description}
\item{I4.t2}{Description}
\item{I5.t2}{Description}
\item{I6.t2}{Description}
\item{I7.t2}{Description}
\item{I8.t2}{Description}
\item{I9.t2}{Description}
\item{I10.t2}{Description}
\item{I11.t2}{Description}
\item{I12.t2}{Description}
\item{I13.t2}{Description}
\item{I14.t2}{Description}
\item{I15.t2}{Description}
}
}
\source{
Adapted from the \pkg{eRm} package.
}
\usage{
lltmdat1
}
\description{
An artificial dataset originally provided with the \pkg{eRm} package,
used for demonstrating Linear Logistic Test Models (LLTM).
}
\keyword{datasets}
