% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2.R
\name{M2}
\alias{M2}
\title{M2 function to test for consistency}
\usage{
M2(dataset, items, method = "gamma")
}
\arguments{
\item{dataset}{A data set with item scores and covariates}

\item{items}{A character vector}

\item{method}{A string determining which associations method to use. Default is
set to gamma, which is Goodman and Kruskal's gamma, other possible methods are
"spearman", "pearson" and "kendall".}
}
\value{
A list with the chosen associations measure, the explicit
associations scores and a status indicating whether or not we passed the
M2 criteria.
}
\description{
M2 function to test for consistency
}
\details{
Missing values are handled using complete cases of the selected item columns.
Missingness in variables not listed in \code{items} does not affect M2.
}
\examples{
data <- toy_spadi_pain
items <- paste0("pain", 1:5)
M2(data, items, method = "gamma")

}
