% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_helper_functions.R
\name{M3_pass}
\alias{M3_pass}
\title{Simple pass/fail flag for M3 function.}
\usage{
M3_pass(M3_output)
}
\arguments{
\item{M3_output}{}
}
\value{
Boolean indicting whether M3 passed or failed
}
\description{
Simple pass/fail flag for M3 function.
}
