% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_LD_helper_functions.R
\name{make_DIF_df}
\alias{make_DIF_df}
\title{DIF list to data frame}
\usage{
make_DIF_df(run_step3c.out)
}
\arguments{
\item{run_step3c.out}{}
}
\value{
A wide format data frame with overwiev of which DIF is deemed
genuin and which is spurious
}
\description{
DIF list to data frame
}
\keyword{internal}
