% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_LD_helper_functions.R
\name{make_source_df}
\alias{make_source_df}
\title{Source list to data frame}
\usage{
make_source_df(run_step3b.out)
}
\arguments{
\item{run_step3b.out}{}
}
\value{
A wide format data frame with overwiev of which DIF is deemed
genuin and which is spurious
}
\description{
Turn the output from run_step3b into a wide format data frame instead of a
list of lsits
}
\keyword{internal}
