% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{partial_gamma_mc_test}
\alias{partial_gamma_mc_test}
\title{Monte Carlo permutation test for partial gamma}
\usage{
partial_gamma_mc_test(dataset, Yi, Xj, strata_vars, B = 1000)
}
\arguments{
\item{dataset}{a data frame dataset}

\item{Yi}{item name}

\item{Xj}{covariate name}

\item{strata_vars}{conditioning variables}

\item{B}{number of simulations}
}
\value{
list with partial gamma coefficient and corrospondoing p_value obtianed
via the Monte Carlo method.
}
\description{
Monte Carlo permutation test for partial gamma
}
\keyword{internal}
