% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{permutation.test}
\alias{permutation.test}
\title{Permutation function for determining p-values from observed partial gamma values.}
\usage{
permutation.test(dataset, Yi, Xj, strata_vars, B = 1000)
}
\arguments{
\item{B}{}
}
\description{
Using the Monte Carlo method (is it Mc tho? - how does MC work?)
}
