% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.item_screening}
\alias{print.item_screening}
\title{print item screening, simply controls what we is printed when we run the
"screen_items" function.}
\usage{
\method{print}{item_screening}(x, ...)
}
\description{
print item screening, simply controls what we is printed when we run the
"screen_items" function.
}
\keyword{internal}
