% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_LD_helper_functions.R
\name{quiet_partgam_DIF}
\alias{quiet_partgam_DIF}
\title{Quiet version of iarm partgam DIF function.}
\usage{
quiet_partgam_DIF(
  dat.items,
  dat.exo,
  p.adj = c("BH", "holm", "hochberg", "hommel", "bonferroni", "BY", "none")
)
}
\arguments{
\item{dat.items}{A data frame or matrix with reposnses to items}

\item{dat.exo}{A data frame or matrix with one or more exogenous grouping
variables.}

\item{p.adj}{The method to be used for adjusting p-values, due to multiple testing}
}
\value{
A data frame with the partial gamma coefficients, standard errors, p-values,
adjusted p-values, confidence limits for every pair of item and exogenous varaible.
}
\description{
Again we needed a function which does exactly what iarm's partgam_DIF function does
but without pringitng the result. As "verbose = FALSE" is not an option we have
taken the function body and commented out the part that prints the results.
}
\keyword{internal}
