% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF_LD_helper_functions.R
\name{quiet_partgam_LD}
\alias{quiet_partgam_LD}
\title{Quiet version of iarm partgam LD function.}
\usage{
quiet_partgam_LD(
  dat.items,
  p.adj = c("BH", "holm", "hochberg", "hommel", "bonferroni", "BY", "none")
)
}
\arguments{
\item{dat.items}{A data frame or matrix of item responses}

\item{p.adj}{A string of charecters defining the methods to be
used for adjusting p-values}
}
\value{
A list with two data frames containing the partial gamma coefficients, standard errors, p-values,
adjusted p-values, confidence limits for every item pair.
}
\description{
This is the exact function that ships with the iarm package, it is called
"partgam_LD". The reason we redefine it is not due to any fancyness, but
simply because the partgam_LD function automatically prints its results and
we want a modified version which does not. Since "verbose = FALSE" is not an
option when calling the function we do it this way. Naturally we found the
function body by running the line "getAnywhere("partgam_LD")".
}
\keyword{internal}
