% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{run_step3b}
\alias{run_step3b}
\title{Apply Step 3(b) to all items}
\usage{
run_step3b(dataset, source_list, items, B = 1000, crit_val = 0.05)
}
\arguments{
\item{dataset}{data.frame}

\item{source_list}{named list of SOURCE(Y_i)}

\item{items}{a charecter vector with the names of items in the dataset}

\item{B}{number of Monte Carlo samples. Default is set to \eqn{1000}.}

\item{crit_val}{significance level. Default is set to \eqn{0.05}.}
}
\value{
list of results per item
}
\description{
Apply Step 3(b) to all items
}
