% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{run_step3c}
\alias{run_step3c}
\title{Apply Step 3(c) to all covariates}
\usage{
run_step3c(
  dataset,
  dif_list,
  items,
  B = 1000,
  crit_val = 0.05,
  p_value_method = c("monte_carlo", "asymptotic", "coin_asymptotic")
)
}
\arguments{
\item{dataset}{data.frame}

\item{dif_list}{named list of DIF(X_j)}

\item{items}{character vector of all item names}

\item{B}{number of Monte Carlo samples}

\item{crit_val}{significance level}

\item{p_value_method}{Character string. Either \code{"monte_carlo"},
\code{"asymptotic"}, or \code{"coin_asymptotic"}.}
}
\value{
list of results per covariate
}
\description{
Apply Step 3(c) to all covariates
}
