% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{s.d_list}
\alias{s.d_list}
\title{Create SOURCE and DIF lists}
\usage{
s.d_list(screen_DIF_output, items, covariates)
}
\arguments{
\item{screen_DIF_output}{The output from screen_DIF data frame frame from the output of
screen_DIF}

\item{items}{A charecter vector containing the item names in the data set}

\item{covariates}{A charecter vector containing the covaraite names in the data set}
}
\value{
a list of lists
\describe{
\item{SOURCE}{A list with the SOURCE varaibles causing DIF for each item}
\item{DIF}{A list with all the DIF items and the exogenous varibales
potentially causing it}
}
}
\description{
We take the output from screen_DIF and create two lists, SOURCE and DIF.
}
