% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{S_n_DIF.list}
\alias{S_n_DIF.list}
\title{Create SOURCE and DIF lists}
\usage{
S_n_DIF.list(reported_DIF)
}
\arguments{
\item{reported_DIF}{The reported_DIF data frame frame from the output of
screen_DIF}
}
\value{
a list of lists
\describe{
\item{SOURCE}{A list with the SOURCE varaibles causing DIF for each item}
\item{DIF}{A list with all the DIF items and the exogenous varibales
potentially causing it}
}
}
\description{
We take the output from screen_DIF and create two lists, SOURCE and DIF.
}
