% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{screen_LD}
\alias{screen_LD}
\title{Local Dependence (LD) Identifier.}
\usage{
screen_LD(df, items, crit_val = 0.05)
}
\arguments{
\item{df}{A data frame containing item responses and covariates.}

\item{items}{A character vector of column names corresponding to the items.}

\item{crit_val}{Numeric. Significance threshold for the adjusted p-values.
Default is set to \code{0.05}.}
}
\value{
A list with two sub-lists, one containing both data frames produced by
\code{iarm::partgam_LD} and one containing the two reduced data frames with only
significant pairs of items. When called directly only the two reduced
data frames are printed.
\describe{
\item{all_LD}{The complete output from \code{iarm::partgam_DIF}.}
\item{reported_LD}{A list with two data frames containing only
item–item pairs with adjusted p-values less than or equal to \code{crit_val}.}
}
}
\description{
LD performs an initial screening for local dependecy between items using the
\code{iarm::partgam_LD} funciton. The function evaluates whether items are
indepndent of each other given their rest score. Note that for each pair of
items 2 such tests can be performed. Take the item pair \eqn{(Y_i,Y_j)} then we
can test both \eqn{Y_i \bot Y_j \mid S-Y_i} and \eqn{Y_j \bot Y_i \mid S-Y_j}.
Hence the screening results in 2 dataframes depending on which item is being
subtracted from the total score \eqn{S}.
}
\details{
P-values are adjusted for multple testing useing the Benjamin-Hochberg procedure.
Item pairs with an adjusted p-value below a specified threshold are flagged as
exhibiting LD.

This function is intended as a screening tool for detecting potential LD.
Flagged item pairs should be investigated further using substantive and
methodological considerations.
}
\examples{
data <- twigg::toy_sp_DIF_and_LD
data_items <- paste0("pain", 1:5)

screen_LD(data, data_items)
}
