% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{step3b.run}
\alias{step3b.run}
\title{Title}
\usage{
step3b.run(
  dataset,
  source.list,
  all_items,
  B = 1000,
  crit_val = 0.05,
  iterative = TRUE
)
}
\arguments{
\item{iterative}{}
}
\description{
Title
}
