% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{step3b_eliminate.sources}
\alias{step3b_eliminate.sources}
\title{Title}
\usage{
step3b_eliminate.sources(
  dataset,
  Yi,
  source.list,
  all_items,
  B = 1000,
  crit_val = 0.05,
  iterative = TRUE
)
}
\arguments{
\item{iterative}{}
}
\description{
Title
}
