% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{step3b_eliminate_sources}
\alias{step3b_eliminate_sources}
\title{Step 3(b): eliminate spurious DIF sources for one item}
\usage{
step3b_eliminate_sources(
  dataset,
  Yi,
  source_set,
  items,
  B = 1000,
  crit_val = 0.05,
  iterative = TRUE
)
}
\arguments{
\item{dataset}{data.frame}

\item{Yi}{item name}

\item{source_set}{candidate covariates}

\item{items}{a charecter vector with the names of items in the dataset}

\item{B}{number of Monte Carlo samples. Default is set to \eqn{1000}.}

\item{crit_val}{significance level. Default is set to \eqn{0.05}.}

\item{iterative}{logical; repeat until stable}
}
\value{
list with remaining_sources and test results
}
\description{
Step 3(b): eliminate spurious DIF sources for one item
}
\keyword{internal}
