% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.R
\name{step4_format_p_value}
\alias{step4_format_p_value}
\title{Format Step 4 Monte Carlo p-values}
\usage{
step4_format_p_value(p, B)
}
\arguments{
\item{p}{Numeric p-values.}

\item{B}{Integer. Number of Monte Carlo samples.}
}
\value{
Character vector of formatted p-values.
}
\description{
Format Step 4 Monte Carlo p-values
}
\keyword{internal}
