% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step6.R
\name{step6_prepare_data}
\alias{step6_prepare_data}
\title{Prepare data for Step 6 conditional independence tests}
\usage{
step6_prepare_data(data, items, covariates, score = NULL)
}
\description{
Prepare data for Step 6 conditional independence tests
}
\keyword{internal}
