% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.R
\name{tidy.gllrm_step4}
\alias{tidy.gllrm_step4}
\title{Tidy a Step 4 structural screen}
\usage{
tidy.gllrm_step4(x, ...)
}
\arguments{
\item{x}{A \code{"gllrm_step4"} object.}

\item{...}{Ignored.}
}
\value{
A data frame containing all Step 4 conditional association tests.
}
\description{
Tidy a Step 4 structural screen
}
