% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_sp_DIF_and_LD.R
\docType{data}
\name{toy_sp_DIF_and_LD}
\alias{toy_sp_DIF_and_LD}
\title{Toy SPADI pain data with both LD and DIF (!)}
\format{
\subsection{toy SPADI pain data}{

A data frame with 250 rows and 7 columns:
\describe{
\item{sex}{Gender of participants. Varaible type: factor with 2 levels "M" male and "F" female.}
\item{age}{Age of participants. Numerical variable ranging from 18 to 85.}
\item{pain1}{ordinal varaible taking values from 0 to 10}
\item{pain2}{ordinal varaible taking values from 0 to 10}
\item{pain3}{ordinal varaible taking values from 0 to 10}
\item{pain4}{ordinal varaible taking values from 0 to 10}
\item{pain5}{ordinal varaible taking values from 0 to 10}
}
}
}
\source{
simulated
}
\usage{
data(toy_sp_DIF_and_LD)
}
\description{
A simulated toy data set for a pain questionnaire to test our functions on.
}
\examples{
data(toy_sp_DIF_and_LD)
summary(toy_sp_DIF_and_LD)
}
\keyword{datasets}
