% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2_helper_functions.R
\name{compute_and_include_total_score}
\alias{compute_and_include_total_score}
\title{Compute and insert total item score}
\usage{
compute_and_include_total_score(data, y)
}
\arguments{
\item{data}{item repose data}

\item{y}{character vector of item names}
}
\value{
data frame of item response with one additional column containing the total score
}
\description{
Compute and insert total item score
}
\examples{
df <- data.frame("id" = c(1,2,3,4), "item1" = c(1,1,1,1), "item2" = c(2,2,2,2))
y <- c("item1", "item2")
compute_and_include_total_score(df, y)
}
