% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_helper_functions.R
\name{cor_one_pair}
\alias{cor_one_pair}
\title{Compute correlation between two variables. Optionally includes the p-value
from the correlation test.}
\usage{
cor_one_pair(x, y, method = "pearson", include_pvalues = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector of same length as x (variable 1)}

\item{method}{Character string for deciding which correlation method to use.}

\item{include_pvalues}{Boolean, indicating whether or not the p-value should
be included. Default is set to TRUE.}
}
\value{
A named list with elements:
\describe{
\item{correlation}{Estimated correlation coefficient.}
\item{p_value}{P-value for the test of zero correlation, or \code{NA}.}
}
}
\description{
Compute correlation between two variables. Optionally includes the p-value
from the correlation test.
}
\examples{
x <- rnorm(50)
y <- x + rnorm(50, sd = 0.1)

cor_one_pair(x, y)
cor_one_pair(x, y, include_pvalues = FALSE)
}
\keyword{internal}
