% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_checks.R
\name{correct.dataset.structure}
\alias{correct.dataset.structure}
\title{convert data set to correct data type}
\usage{
correct.dataset.structure(dataset)
}
\arguments{
\item{dataset}{a dataset, which can be a matrix, data frame of tibble like structure.}
}
\value{
the dataset but as a data frame instead of whatever class it was origional
}
\description{
convert data set to correct data type
}
