% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m1.R
\name{M1}
\alias{M1}
\title{M1 test for consistence}
\usage{
M1(dataset, items)
}
\arguments{
\item{dataset}{A data set with item scores and covariates}

\item{y}{A character vector with item names}
}
\value{
A Boolean indicating whether the dataset is consistent wrt the M1
criteria
}
\description{
M1 test for consistence
}
\examples{
df <- toy_spadi_pain
items <- paste0("pain", 1:5)
M1(df, items)

}
