% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m1.R
\name{M1}
\alias{M1}
\title{M1 test for consitence}
\usage{
M1(dataset, y = vector_with_names_of_items)
}
\arguments{
\item{dataset}{A data set with item scores and covariates}

\item{y}{A character vector}
}
\value{
A Boolean indicating whether the dataset is consistent wrt the M1
criteria
}
\description{
M1 test for consitence
}
\examples{
df <- data
y <- c("item1", "item2", "item3")
M1(df, y)

}
