% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2.R
\name{M2}
\alias{M2}
\title{M2 function to test for consistency}
\usage{
M2(dataset, y = vector_with_names_of_items)
}
\arguments{
\item{dataset}{A data set with item scores and covariates}

\item{y}{A character vector}
}
\value{
A Boolean indicating wheter the data set is consitent or not
}
\description{
M2 function to test for consistency
}
