% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2.R
\name{M2}
\alias{M2}
\title{M2 function to test for consistency}
\usage{
M2(dataset, items)
}
\arguments{
\item{dataset}{A data set with item scores and covariates}

\item{items}{A character vector}
}
\value{
A Boolean indicating whether the data set is consistent or not
}
\description{
M2 function to test for consistency
}
