% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3.R
\name{M3}
\alias{M3}
\title{Title}
\usage{
M3(
  dataset,
  y = vector_with_names_of_items,
  X = vector_with_names_of_covariates,
  corr_method = "pearson",
  include_pvalues = TRUE
)
}
\arguments{
\item{include_pvalues}{}
}
\description{
Title
}
