% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{partial_gamma_coin_test}
\alias{partial_gamma_coin_test}
\title{Conditional Monte Carlo test for DIF using coin}
\usage{
partial_gamma_coin_test(dataset, Yi, Xj, strata_vars, B = 10000)
}
\arguments{
\item{dataset}{data.frame}

\item{Yi}{item name}

\item{Xj}{covariate name}

\item{strata_vars}{conditioning variables}

\item{B}{Monte Carlo iterations}
}
\value{
list with gamma and p_value
}
\description{
Conditional Monte Carlo test for DIF using coin
}
\keyword{internal}
