% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_items.R
\name{screen_items}
\alias{screen_items}
\title{Run item screening procedure}
\usage{
screen_items(
  dataset,
  items,
  covariates = NULL,
  method = "gamma",
  run_M1 = TRUE,
  run_M2 = TRUE,
  run_M3 = TRUE
)
}
\arguments{
\item{dataset}{A data frame containing item responses and covariates.}

\item{items}{Character vector giving the names of item response variables.}

\item{covariates}{Optional character vector giving the names of covariates
to be screened using the M3 criterion.}

\item{method}{Character string specifying the desired measure of association}

\item{run_M1}{Logical; should the M1 screening step be performed?}

\item{run_M2}{Logical; should the M2 screening step be performed?}

\item{run_M3}{Logical; should the M3 screening step be performed?}
}
\value{
An object of class \code{"item_screening"}, which is a named list
containing:
\describe{
\item{passed}{Logical; \code{TRUE} if all screening steps passed.}
\item{failed_step}{Character string indicating the step at which screening
failed, or \code{NULL} if all steps passed.}
\item{M1}{Result of the M1 screening step (if performed).}
\item{M2}{Result of the M2 screening step (if performed).}
\item{M3}{Data frame with M3 correlation results (if performed).}
}
}
\description{
Runs the item screening procedure by sequentially applying consistency checks
M1, M2, and M3 as proposed in the item screening framework. The procedure
does NOT stop early if a screening step fails. The final print contains results
from all the steps (M1, M2, and M3) and prints explicitly if we passed all steps or
not. If we did not pass all steps, the name of the failed steps are printed
explicitly. However, we can manually decide whether or not individual steps
should be performed, default is that all steps are performed, but can be changed
by passing the argument "run_M1 = FALSE" e.g. However M3 is only run
if "run_M3=TRUE" (default) and some covariates are passed to the function
(otherwise it does not make sense to run it hence it wont).
}
\examples{
items <- paste0("pain", 1:5)

res <- screen_items(
  dataset = toy_spadi_pain,
  items = items,
  covariates = c("age", "sex")
)

res$passed
res$failed_step
}
